requirements_3.6.12.txt
scikit-learn==0.19.1
wget==3.2
tensorflow-cpu==2.1.2
SpeechRecognition==3.8.1
PyAudio==0.2.11
opencv-python==4.4.0.44
numpy==1.18.5
nltk==3.5
keyboard==0.13.5
Flask==1.1.2
dlib==19.7.0
scipy==1.5.4
h5py==2.10.0
